First Connection with backend
