+++ lib/properties-searching/ui/post-ui/see_post.dart
