+++ lib/properties-searching/ui/post-ui/showPosts.dart
@@ -1,0 +1,221 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:renstatefrontend/properties-searching/ui/post-ui/see_post.dart';
+import 'package:renstatefrontend/shared/appBarApp.dart';
+import 'package:renstatefrontend/shared/bottomNavigationApp.dart';
+
+import '../../../models/post.dart';
+import '../../../shared/services/PostService.dart';
+
+
+class ShowPosts extends StatefulWidget {
+  const ShowPosts({super.key});
+
+  @override
+  State<ShowPosts> createState() => _ShowPostsState();
+}
+
+class _ShowPostsState extends State<ShowPosts> {
+
+  final postService = PostService();
+  late Future<List<Post>> _posts;
+
+  @override
+  void initState() {
+    super.initState();
+    _posts = postService.getPosts();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      resizeToAvoidBottomInset: false,
+      body: FutureBuilder(
+          future: _posts,
+          builder: (context, snapshot){
+            if (snapshot.connectionState == ConnectionState.waiting) {
+              return Center(child: CircularProgressIndicator());
+            } else if (snapshot.hasError) {
+              return Center(child: Text('Error: ${snapshot.error}'));
+            } else if (!snapshot.hasData || snapshot.data!.isEmpty) {
+              return Center(child: Text('No posts available.'));
+            }else{
+              return
+                Center(
+                child: ListView(
+                  children: _listPosts(snapshot.data!),
+                ),
+              );
+            }
+          },
+      ),
+      bottomNavigationBar: bottomNavigationApp(context),
+    );
+  }
+
+  List<Widget> _listPosts(List<Post>data){
+    List<Widget> posts = [];
+    posts.add(titleResult());
+    posts.add(searchDesign());
+    for (var post in data){
+      posts.add(
+        Center(
+          child: ListView(
+            physics: NeverScrollableScrollPhysics(),
+            shrinkWrap: true,
+            children: [
+              Padding(
+                padding: const EdgeInsets.symmetric(vertical: 15.0),
+                child: viewPost(context, post),
+              ),
+            ],
+          ),
+        ),
+      );
+    }
+    return posts;
+  }
+}
+
+Widget viewPost(BuildContext context, Post post) {
+  return FractionallySizedBox(
+    widthFactor: 0.9,
+    child: Card(
+      color: Color(0xFF064789),
+      elevation: 5,
+      child: Padding(
+        padding: const EdgeInsets.all(10.0),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.center,
+            children: [
+              titleCard(post),
+              SizedBox(height: 15.0),
+              FractionallySizedBox(
+                widthFactor: 0.5,
+                child: Image(
+                  image: NetworkImage(post.imgUrl),
+                  fit: BoxFit.cover,
+                ),
+              ),
+              SizedBox(height: 15.0),
+              Text(
+                post.description,
+                softWrap: true,
+                style: TextStyle(
+                  color: Colors.white,
+                  fontWeight: FontWeight.bold,
+                  fontSize: 18.0,
+                ),
+                textAlign: TextAlign.center,
+              ),
+              SizedBox(height: 30.0,),
+              buttonDetails(context)
+            ],
+          ),
+      ),
+    ),
+  );
+}
+
+
+Widget titleCard(Post post){
+  return Row(
+    children: [
+      Expanded(
+        child: Text(
+          post.title,
+          style: TextStyle(
+            fontSize: 23.0,
+            fontWeight: FontWeight.bold,
+            color: Colors.white,
+          ),
+          textAlign: TextAlign.center,
+        ),
+      ),
+      Container(
+        padding: EdgeInsets.symmetric(vertical: 5.0, horizontal: 15.0),
+        decoration: BoxDecoration(
+          color: Colors.white,
+          borderRadius: BorderRadius.circular(8.0),
+        ),
+        child: Text(
+          'S/ ' + post.price.toString(),
+          style: TextStyle(
+            fontSize: 20.0,
+            fontWeight: FontWeight.bold,
+            color: Colors.black,
+          ),
+        ),
+      ),
+    ],
+  );
+}
+
+Widget searchDesign() {
+  return FractionallySizedBox(
+    widthFactor: 0.9,
+    child: Column(
+      children: [
+        Text(
+          'Search',
+          style: TextStyle(
+            color: Color(0xFF064789),
+            fontSize: 36,
+            fontWeight: FontWeight.w400,
+          ),
+        ),
+        TextField(
+          style: TextStyle(fontFamily: 'Inter'),
+          maxLines: 1,
+          decoration: InputDecoration(
+            border: OutlineInputBorder(),
+            suffixIcon: Icon(Icons.search),
+          ),
+        ),
+      ],
+    ),
+  );
+}
+
+Widget titleResult() {
+  return Padding(
+    padding: const EdgeInsets.symmetric(vertical: 15.0),
+    child: Text(
+      'ALL RESULTS',
+      style: TextStyle(
+        color: Colors.black,
+        fontSize: 20.0,
+        fontFamily: 'Roboto',
+        fontWeight: FontWeight.w900,
+        letterSpacing: 0.52,
+      ),
+      textAlign: TextAlign.center,
+    ),
+  );
+}
+
+Widget buttonDetails(context) {
+  return Container(
+    decoration: BoxDecoration(
+      color: Colors.blue,
+      borderRadius: BorderRadius.circular(5),
+    ),
+    child: TextButton(
+      child: Text(
+        'Detalles',
+        style: TextStyle(
+          color: Colors.white,
+          fontSize: 18,
+          fontFamily: 'Inter',
+          fontWeight: FontWeight.w700,
+          letterSpacing: 0.52,
+        ),
+      ),
+      onPressed: () {
+        Navigator.push(context,
+            MaterialPageRoute(builder: (context)=>SeePost())
+        );
+      },
+    ),
+  );
+}
